chore(license): migrate from BUSL-1.1 to AGPL-3.0 with commercial ee/ split
- Replace root LICENSE with AGPL-3.0-only plus ee/ carve-out preamble;
  prior BUSL releases keep their original terms
- Move cloud-only onboarding module to packages/backend/src/ee/cloud
  under the AnythingMCP Commercial License
- Add CLA.md and contributor-assistant workflow gating PRs on signature
- Update license metadata and docs: package.json fields, CITATION.cff,
  glama.json, README, CONTRIBUTING, SUPPORT, license FAQ
