+++ CITATION.cff
@@ -30,4 +30,4 @@ keywords:
   - soap-to-mcp
   - graphql-to-mcp
   - database-to-mcp
-license: BUSL-1.1
+license: AGPL-3.0-only
