+++ CLA.md
@@ -1,0 +1,76 @@
+# AnythingMCP Individual Contributor License Agreement
+
+Thank you for your interest in contributing to AnythingMCP, a project
+of helpcode.ai GmbH ("we", "us"). This Contributor License Agreement
+("CLA") clarifies the intellectual-property rights granted with your
+contributions. It protects you, us, and the users of the project; it
+does not change your right to use your own contributions for any other
+purpose.
+
+By signing this CLA (via the automated check on your pull request),
+you accept the following terms for your present and future
+contributions submitted to AnythingMCP.
+
+## 1. Definitions
+
+**"Contribution"** means any original work of authorship, including
+any modifications or additions to an existing work, that you
+intentionally submit to us for inclusion in AnythingMCP, in any form
+(source code, documentation, configuration, etc.).
+
+**"You"** means the individual who submits a Contribution, or the
+legal entity on whose behalf it is submitted.
+
+## 2. Grant of copyright license
+
+You grant helpcode.ai GmbH a perpetual, worldwide, non-exclusive,
+royalty-free, irrevocable copyright license to reproduce, prepare
+derivative works of, publicly display, publicly perform, sublicense,
+and distribute your Contributions and derivative works thereof, and to
+**license and re-license your Contributions under any license terms**,
+including open-source licenses (such as the AGPL-3.0) and commercial
+licenses.
+
+## 3. Grant of patent license
+
+You grant helpcode.ai GmbH and recipients of software distributed by
+us a perpetual, worldwide, non-exclusive, royalty-free, irrevocable
+(except as stated below) patent license to make, have made, use, offer
+to sell, sell, import, and otherwise transfer your Contributions,
+where such license applies only to those patent claims licensable by
+you that are necessarily infringed by your Contribution alone or in
+combination with the work to which it was submitted. If any entity
+institutes patent litigation alleging that your Contribution or the
+work infringes a patent, any patent licenses granted to that entity
+under this CLA for that Contribution or work terminate as of the date
+such litigation is filed.
+
+## 4. Your representations
+
+You represent that:
+
+1. You are legally entitled to grant the above licenses. If your
+   employer has rights to intellectual property you create, you have
+   received permission to make the Contribution on its behalf, or your
+   employer has waived such rights.
+2. Each Contribution is your original creation, or you have identified
+   the parts that are not, together with their license and origin.
+3. Your Contribution does not knowingly infringe any third party's
+   intellectual-property rights.
+
+## 5. You keep your rights
+
+You retain ownership of the copyright in your Contributions. This CLA
+is a license, not a copyright assignment. You remain free to use,
+publish, and license your own Contributions however you wish.
+
+## 6. No obligation and no warranty
+
+We are not obligated to include your Contribution in the project.
+Unless required by applicable law or agreed in writing, you provide
+your Contributions "AS IS", without warranties or conditions of any
+kind.
+
+---
+
+Questions? Contact licensing@helpcode.ai.
