+++ SUPPORT.md
@@ -19,7 +19,7 @@ The fastest answers usually come from other users. Before opening an issue, chec
 - **[Deployment Guide](docs/deployment.md)** — production self-hosting
 - **[API Reference](docs/api-reference.md)** — full REST API
 - **[Tool Definition Format](docs/tool-definition.md)** — how adapter JSON files are structured
-- **[License FAQ](docs/license-faq.md)** — what BSL-1.1 means in practice
+- **[License FAQ](docs/license-faq.md)** — what the AGPL-3.0 license means in practice
 
 ## Bug reports
