+++ packages/backend/src/app.module.ts
@@ -33,7 +33,7 @@ import { OAuthUrlRewriteInterceptor } from './auth/oauth-url-rewrite.interceptor
 import { EmailVerifiedGuard } from './auth/email-verified.guard';
 import { AdaptersModule } from './adapters/adapters.module';
 import { OrganizationsModule } from './organizations/organizations.module';
-import { CloudModule } from './cloud/cloud.module';
+import { CloudModule } from './ee/cloud/cloud.module';
 import { getRequiredSecret } from './common/secrets.util';
 import { AppLoggerModule } from './common/logger.module';
