+++ packages/frontend/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.33",
   "description": "AnythingMCP — Next.js Admin UI",
   "private": true,
-  "license": "BUSL-1.1",
+  "license": "AGPL-3.0-only",
   "scripts": {
     "dev": "next dev --port 3000",
     "build": "next build",
